Fix permissions configuration bug
When creating permissions without specifying any commands,
the output ClientPermissions.xml is not formatted correctly
and cannot be loaded by the game.
Ensure the XML tag is closed correctly when commands are not present.
Update comments and README to reflect correct usage of separators in
BAR_PERMISSIONS environment variable and provide more complete context
around usage.